requirements.txt
numpy==1.16.2
matplotlib==3.0.3
opencv_python==4.1.0.25
torch==1.1.0
pandas==0.24.2
torchvision==0.3.0
Pillow==6.2.0
imageio==2.6.1
